Add messaging, validation, and user services
- Integrated RabbitMQ with MassTransit for event-driven messaging.
- Added FluentValidation for request validation in Basket.Api.
- Introduced BasketCheckoutHandler and related event handlers.
- Created BuildingBlocks.Messaging for shared messaging logic.
- Added Users.api microservice with MongoDB and JWT authentication.
- Implemented Argon2Hasher for secure password hashing.
- Introduced YARPGateway.api as a reverse proxy with rate limiting.
- Updated Docker and solution files to include new services.

+++ Basket.Api/Features/Commands/CreateBasket/CreateBasketHandler.cs
@@ -1,4 +1,5 @@
-﻿using Basket.Api.Data.Entities;
+﻿using Basket.Api.Application.Exceptions;
+using Basket.Api.Data.Entities;
 using Basket.Api.Dtos;
 using Basket.Api.Repositories;
 using Discount.gRPC.Protos;
@@ -21,7 +22,10 @@ namespace Basket.Api.Features.Commands.CreateBasket
         public async Task<CreateBasketResponse> Handle(CreateBasketCommand request, CancellationToken cancellationToken)
         {
             var existingBasket =await _repository.GetShoppingCartAsync(request.Username);
-            if (existingBasket != null) { 
+            if (existingBasket != null) {
+                if (existingBasket.CheckoutStatus != CheckoutStatus.None) {
+                    throw new BadRequestException("basket is already being proccessed");
+                }
                await _repository.DeleteShoppingCartAsync(request.Username);
             }
             await CalculateDiscountAmounts(request.Items);
